Split Lang.Crucible.JVM.Translation into two modules.
`Lang.Crucible.JVM.Translation` now only exports the functions
to translate a single JVM method. Other functionality related
to setting up the symbolic simulator has moved to the new
module `Lang.Crucible.JVM.Simulate`.

Also, the module re-exports from `Lang.Crucible.JVM.Translation`
have been removed. To make up for this, a new top-level module
`Lang.Crucible.JVM` has been added, which re-exports the other
modules below it in the module hierarchy.
