+++ crucible-jvm/tool/Main.hs
@@ -77,7 +77,7 @@ import qualified Crux.Options  as Crux
 
 import qualified Lang.JVM.Codebase as JCB
 
-import           Lang.Crucible.JVM.Translation
+import           Lang.Crucible.JVM.Simulate (executeCrucibleJVMCrux)
 import           Lang.Crucible.JVM.Types
 
 -- executable
